FSAi. Actualizacion a FW2.8.

+++ FSMail/Imap.cs
@@ -73,7 +73,7 @@ namespace FSMail
 		/// <param name="sslEnabled"> </param>
 		/// <exception cref="IMAP_ERR_LOGIN"
 		/// <exception cref="IMAP_ERR_INVALIDPARAM"
-		public void Login(string sHost, ushort nPort, string sUserId, string sPassword, bool sslEnabled = false)
+		public void Login(string sHost, int nPort, string sUserId, string sPassword, bool sslEnabled = false)
 		{
 			ImapResponseEnum eImapResponse = ImapResponseEnum.IMAP_SUCCESS_RESPONSE;
 			ImapException e_login = new ImapException(ImapException.ImapErrorEnum.IMAP_ERR_LOGIN, m_sUserId);
